Generic compression and result
