+++ Project.toml
@@ -15,6 +15,7 @@ SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 CliqueTrees = "60701a23-6482-424a-84db-faee86b9b1f8"
 Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
+GPUArrays = "0c68f7d7-f131-5f86-a1c3-88cf8149b2d7"
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
 MathOptInterface = "b8f27783-ece8-5eb3-8dc8-9495eed66fee"
 cuSPARSE = "b26da814-b3bc-49ef-b0ee-c816305aa060"
@@ -23,19 +24,21 @@ cuSPARSE = "b26da814-b3bc-49ef-b0ee-c816305aa060"
 SparseMatrixColoringsCUDAExt = ["CUDA", "cuSPARSE"]
 SparseMatrixColoringsCliqueTreesExt = "CliqueTrees"
 SparseMatrixColoringsColorsExt = "Colors"
+SparseMatrixColoringsGPUArraysExt = "GPUArrays"
 SparseMatrixColoringsJuMPExt = ["JuMP", "MathOptInterface"]
 
 [compat]
 ADTypes = "1.2.1"
 CUDA = "6.0.0"
-cuSPARSE = "6.0.0"
 CliqueTrees = "1"
 Colors = "0.12.11, 0.13"
 DocStringExtensions = "0.8,0.9"
+GPUArrays = "11.5.0"
 JuMP = "1.29.1"
 LinearAlgebra = "1"
 MathOptInterface = "1.45.0"
 PrecompileTools = "1.2.1"
 Random = "1"
 SparseArrays = "1"
+cuSPARSE = "6.0.0"
 julia = "1.10"
